Consolidate down to one header file
The other two header files were very small, so let's keep it at one
while we can.
