--- include/fgt/exceptions.hpp
@@ -1,36 +1,0 @@
-// fgt, C++ library for Fast Gauss Transforms
-// Copyright (C) 2015 Peter J. Gadomski <pete.gadomski@gmail.com>
-//
-// This library is free software; you can redistribute it and/or modify it
-// under the terms of the GNU Lesser General Public License as published by the
-// Free Software Foundation; either version 2.1 of the License, or (at your
-// option) any later version.
-//
-// This library is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
-// for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with this library; if not, write to the Free Software Foundation,
-// Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
-
-#pragma once
-
-#include <stdexcept>
-
-
-namespace fgt {
-
-
-class fgt_error : public std::runtime_error {
-public:
-    using std::runtime_error::runtime_error;
-};
-
-
-class dimension_mismatch : public fgt_error {
-public:
-    using fgt_error::fgt_error;
-};
-}
