+++ src/gauss_transform.cpp
@@ -15,9 +15,7 @@
 // along with this library; if not, write to the Free Software Foundation,
 // Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
-#include <fgt/fgt.hpp>
-
-#include <fgt/exceptions.hpp>
+#include "fgt.hpp"
 
 
 namespace fgt {
