+++ src/ifgt.cpp
@@ -15,9 +15,8 @@
 // along with this library; if not, write to the Free Software Foundation,
 // Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
-#include <fgt/fgt.hpp>
-
 #include "clustering.hpp"
+#include "fgt.hpp"
 #include "monomials.hpp"
 #include "nchoosek.hpp"
 #include "p_max_total.hpp"
