+++ test/ifgt_test.cpp
@@ -15,14 +15,12 @@
 // along with this library; if not, write to the Free Software Foundation,
 // Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
-#include <fgt/fgt.hpp>
-
-#include "clustering.hpp"
-#include "config.hpp"
-
 #include <armadillo>
 #include "gtest/gtest.h"
 
+#include "clustering.hpp"
+#include "fgt.hpp"
+#include "config.hpp"
 
 namespace fgt {
