+++ test/trial.cpp
@@ -15,10 +15,10 @@
 // along with this library; if not, write to the Free Software Foundation,
 // Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
-#include <fgt/fgt.hpp>
-
 #include <armadillo>
 #include "gtest/gtest.h"
+
+#include "fgt.hpp"
 
 
 namespace fgt {
